refactor: rename ptm/test.py to ptm/lmm.py (PR-3) (#168)
algorithms/ptm/test.py holds the LMM constraint-test functions (run_lmm,
run_binned_interaction_lmm, LMMResult, BinnedLMMResult). A module literally
named test.py is confusing structure for a stats module and invites import
ambiguity. Rename to the descriptive lmm.py and update all references:
- git mv test.py -> lmm.py (plus the docstring self-reference)
- ptm/__init__.py lazy-import map (4 entries)
- tools/ptm/ptm_cli.py (2 imports) and algorithms/ptm/report.py (1 import)
- stale module references in ptm/constants.py comments

No behavior change. flake8 clean; the lazy symbols (run_lmm etc.) still
resolve via hvantk.algorithms.ptm on the project venv (3.10).
